+++ src/main/java/org/opentripplanner/common/geometry/CompactLineString.java
@@ -1,7 +1,5 @@
 package org.opentripplanner.common.geometry;
 
-import java.io.Serializable;
-
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.GeometryFactory;
 import org.locationtech.jts.geom.LineString;
@@ -24,9 +22,7 @@ import org.locationtech.jts.geom.LineString;
  * 
  * @author laurent
  */
-public final class CompactLineString implements Serializable {
-
-    private static final long serialVersionUID = 1L;
+public final class CompactLineString {
 
     /**
      * Multiplier for fixed-float representation. For lat/lon CRS, 1e6 leads to a precision of 0.11
@@ -44,9 +40,7 @@ public final class CompactLineString implements Serializable {
     /**
      * Singleton representation of a straight-line (where nothing has to be stored), to be re-used.
      */
-    protected static final int[] STRAIGHT_LINE = new int[0];
-
-    protected static final byte[] STRAIGHT_LINE_PACKED = new byte[0];
+    public static final byte[] STRAIGHT_LINE_PACKED = new byte[0];
 
     /**
      * Geometry factory. TODO - Do we need to make this parametrable?
@@ -66,12 +60,12 @@ public final class CompactLineString implements Serializable {
      * @param reverse True if A and B are inverted (B is start, A is end).
      * @return
      */
-    public static int[] compactLineString(double xa, double ya, double xb, double yb,
+    public static byte[] compactLineString(double xa, double ya, double xb, double yb,
             LineString lineString, boolean reverse) {
         if (lineString == null)
             return null;
         if (lineString.getCoordinates().length == 2)
-            return STRAIGHT_LINE;
+            return STRAIGHT_LINE_PACKED;
         double x0 = reverse ? xb : xa;
         double y0 = reverse ? yb : ya;
         double x1 = reverse ? xa : xb;
@@ -102,40 +96,34 @@ public final class CompactLineString implements Serializable {
             oix = ix;
             oiy = iy;
         }
-        return coords;
-    }
-
-    /**
-     * Same as the other version, but in a var-len int packed form (Dlugosz coding).
-     * 
-     * @param x0
-     * @param y0
-     * @param x1
-     * @param y1
-     * @param lineString
-     * @return
-     */
-    public static byte[] compackLineString(double x0, double y0, double x1, double y1,
-            LineString lineString, boolean reverse) {
-        int[] coords = compactLineString(x0, y0, x1, y1, lineString, reverse);
-        if (coords == STRAIGHT_LINE)
-            return STRAIGHT_LINE_PACKED;
         return DlugoszVarLenIntPacker.pack(coords);
     }
 
     /**
-     * Construct a LineString based on external end-points and compacted int version.
-     * 
-     * @param xa
-     * @param ya
-     * @param xb
-     * @param yb
-     * @param coords Compact version of coordinates
-     * @param reverse True if A and B and the compacted geometry is reversed.
+     * Wrapper for the above method in the case where there are no start/end coordinates provided.
+     * 0-coordinates are added in order for the delta encoding to work correctly.
+     */
+    public static byte[] compactLineString(LineString lineString, boolean reverse) {
+        lineString = GeometryUtils.addStartEndCoordinatesToLineString(
+                new Coordinate(0.0, 0.0),
+                lineString,
+                new Coordinate(0.0, 0.0));
+        return compactLineString(0.0, 0.0, 0.0, 0.0, lineString, reverse);
+    }
+
+    /**
+     * Same as the other version, but in a var-len int packed form (Dlugosz coding).
+     *
+     * @param xa X coordinate of end point A
+     * @param ya Y coordinate of end point A
+     * @param xb X coordinate of end point B
+     * @param yb Y coordinate of end point B
+     * @param packedCoords The byte array to uncompact
      * @return
      */
     public static LineString uncompactLineString(double xa, double ya, double xb, double yb,
-            int[] coords, boolean reverse) {
+            byte[] packedCoords, boolean reverse) {
+        int[] coords = DlugoszVarLenIntPacker.unpack(packedCoords);
         int size = coords == null ? 2 : (coords.length / 2) + 2;
         Coordinate[] c = new Coordinate[size];
         double x0 = reverse ? xb : xa;
@@ -162,17 +150,11 @@ public final class CompactLineString implements Serializable {
     }
 
     /**
-     * Same as the other version, but in a var-len int packed form (Dlugosz coding).
-     * 
-     * @param x0
-     * @param y0
-     * @param x1
-     * @param y1
-     * @param coords
-     * @return
+     * Wrapper for the above method in the case where there are no start/end coordinates provided.
+     * 0-coordinates are added and then removed in order for the delta encoding to work correctly.
      */
-    public static LineString uncompackLineString(double x0, double y0, double x1, double y1,
-            byte[] packedCoords, boolean reverse) {
-        return uncompactLineString(x0, y0, x1, y1, DlugoszVarLenIntPacker.unpack(packedCoords), reverse);
+    public static LineString uncompactLineString(byte[] packedCoords, boolean reverse) {
+        LineString lineString = uncompactLineString(0.0, 0.0, 0.0, 0.0, packedCoords, reverse);
+        return GeometryUtils.removeStartEndCoordinatesFromLineString(lineString);
     }
 }