+++ src/main/java/org/opentripplanner/index/IndexAPI.java
@@ -530,7 +530,7 @@ public class IndexAPI {
     public Response getGeometryForPattern (@PathParam("patternId") String patternIdString) {
         TripPattern pattern = index.graph.tripPatternForId.get(patternIdString);
         if (pattern != null) {
-            EncodedPolylineBean geometry = PolylineEncoder.createEncodings(pattern.geometry);
+            EncodedPolylineBean geometry = PolylineEncoder.createEncodings(pattern.getGeometry());
             return Response.status(Status.OK).entity(geometry).build();
         } else {
             return Response.status(Status.NOT_FOUND).entity(MSG_404).build();