+++ src/main/java/org/opentripplanner/routing/algorithm/raptor/router/street/AccessEgressRouter.java
@@ -42,7 +42,6 @@ public class AccessEgressRouter {
                     new Transfer(-1,
                             (int)stopAtDistance.edges.stream().map(Edge::getDistance)
                                     .collect(Collectors.summarizingDouble(Double::doubleValue)).getSum(),
-                            Arrays.asList(stopAtDistance.geom.getCoordinates()),
                             stopAtDistance.edges));
         }