+++ src/main/java/org/opentripplanner/routing/algorithm/raptor/transit/Transfer.java
@@ -3,6 +3,8 @@ package org.opentripplanner.routing.algorithm.raptor.transit;
 import org.locationtech.jts.geom.Coordinate;
 import org.opentripplanner.routing.graph.Edge;
 
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
@@ -11,25 +13,27 @@ public class Transfer {
 
     private final int distanceMeters; // TODO Add units in the name of the field
 
-    private final List<Coordinate> coordinates;
-
     private final List<Edge> edges;
 
-    public Transfer(int toStop, int distanceMeters, List<Coordinate> coordinates) {
+    public Transfer(int toStop, int distanceMeters) {
         this.toStop = toStop;
         this.distanceMeters = distanceMeters;
-        this.coordinates = coordinates;
         this.edges = Collections.emptyList();
     }
 
-    public Transfer(int toStop, int distanceMeters, List<Coordinate> coordinates, List<Edge> edges) {
+    public Transfer(int toStop, int distanceMeters, List<Edge> edges) {
         this.toStop = toStop;
         this.distanceMeters = distanceMeters;
-        this.coordinates = coordinates;
         this.edges = edges;
     }
 
     public List<Coordinate> getCoordinates() {
+        List<Coordinate> coordinates = new ArrayList<>();
+        for (Edge edge : edges) {
+            if (edge.getGeometry() != null) {
+                coordinates.addAll((Arrays.asList(edge.getGeometry().getCoordinates())));
+            }
+        }
         return coordinates;
     }